+++ src/dcos_e2e_cli/dcos_docker/commands/mac_network.py
@@ -20,14 +20,10 @@ from dcos_e2e.backends import Docker
 
 from ._common import docker_client
 
-# We start these names with "vpn" to avoid a conflict with
-# "minidcos docker clean".
-_PROXY_CONTAINER_NAME = 'vpn-{container_name_prefix}-proxy'.format(
-    container_name_prefix=Docker().container_name_prefix,
-)
-_OPENVPN_CONTAINER_NAME = 'vpn-{container_name_prefix}-openvpn'.format(
-    container_name_prefix=Docker().container_name_prefix,
-)
+# These names cannot include the standard container name prefix else they
+# conflict with "minidcos docker clean".
+_PROXY_CONTAINER_NAME = 'vpn-proxy'
+_OPENVPN_CONTAINER_NAME = 'vpn-openvpn'
 
 
 def _validate_ovpn_file_does_not_exist(