+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Next
 ----
 
 * Remove use of ``select`` which is not supported on Windows.
+* ``minidcos docker clean`` will no longer clean up containers which are started from now on by the tooling to create a custom macOS network.
 
 2019.04.25.0
 ------------