+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 Next
 ----
 
+2019.04.29.0
+------------
+
 * Remove use of ``select`` which is not supported on Windows.
 * ``minidcos docker clean`` will no longer clean up containers which are started from now on by the tooling to create a custom macOS network.